requirements.txt
numpy==1.18.4
pytorch==1.5.0
tokenizers==0.8.1
transformers==3.0.2